Refactoring dynamic_review_assignment_helper.rb for find_submissions_in_current_cycle to return teams
The find_submissions_in_current_cycle method returned a list of
participants.  It will now return a list of teams.  Returning a list of
participants leads to redundant entries in the hash, if any are done by
multi-member teams.  There used to be a need for dealing with
participants rather than teams, but in the current Expertiza,
submissions always belong to teams rather than individuals